test: consolidate touch target checks
